+++ example/src/main/scala/tamer/example/DatabaseGeneralized.scala
@@ -1,37 +1,18 @@
-package tamer
-package example
+package tamer.example
 
-import tamer.config.Config
-import tamer.db.Db.Timestamped
-import tamer.db.{ConfigDb, DbTransactor, HashableState, InstantOps, QueryResult, TamerDBConfig}
-import tamer.kafka.Kafka
-import zio.blocking.Blocking
-import zio._
 import doobie.implicits.legacy.instant._
 import doobie.syntax.string._
+import tamer.{TamerError, db}
+import tamer.config.Config
 import tamer.db.ConfigDb.{DbConfig, QueryConfig}
+import tamer.db.{ConfigDb, DbTransactor, HashableState, InstantOps, QueryResult, Setup, TamerDBConfig}
+import tamer.kafka.Kafka
+import zio._
+import zio.blocking.Blocking
 
+import java.time.Instant
 import java.time.temporal.ChronoUnit._
-import java.time.{Duration, Instant}
 import scala.util.hashing.byteswap64
-
-final case class Key(id: String)
-final case class Value(id: String, name: String, description: Option[String], modifiedAt: Instant) extends Timestamped(modifiedAt)
-
-object Main extends zio.App {
-  val program: ZIO[ZEnv, TamerError, Unit] = (for {
-    boot <- UIO(Instant.now())
-    _ <- tamer.db.fetchWithTimeSegment(ts =>
-      sql"""SELECT id, name, description, modified_at FROM users WHERE modified_at > ${ts.from} AND modified_at <= ${ts.to}""".query[Value]
-    )(
-      earliest = boot.minus(60, DAYS),
-      tumblingStep = Duration.of(5, MINUTES),
-      keyExtract = (value: Value) => Key(value.id)
-    )
-  } yield ()).mapError(e => TamerError("Could not run tamer example", e))
-
-  override final def run(args: List[String]): URIO[ZEnv, ExitCode] = program.exitCode
-}
 
 final case class MyState(from: Instant, to: Instant) extends HashableState {
 
@@ -42,7 +23,7 @@ final case class MyState(from: Instant, to: Instant) extends HashableState {
   override val stateHash: Int = (byteswap64(from.getEpochSecond) + byteswap64(to.getEpochSecond)).intValue
 }
 
-object MainGeneralized extends zio.App {
+object DatabaseGeneralized extends zio.App {
   val transactorLayer: Layer[TamerError, DbTransactor]                     = (Blocking.live ++ ConfigDb.live) >>> db.hikariLayer
   val kafkaLayer: Layer[TamerError, Kafka]                                 = Config.live >>> Kafka.live
   val queryConfigLayer: Layer[TamerError, DbConfig with QueryConfig]       = ConfigDb.live
@@ -50,7 +31,7 @@ object MainGeneralized extends zio.App {
   val program: ZIO[Kafka with TamerDBConfig with ZEnv, TamerError, Unit] = (for {
     boot <- UIO(Instant.now())
     earliest = boot.minus(60, DAYS)
-    setup = tamer.db.mkSetup((s: MyState) =>
+    setup = Setup((s: MyState) =>
       sql"""SELECT id, name, description, modified_at FROM users WHERE modified_at > ${s.from} AND modified_at <= ${s.to}""".query[Value]
     )(
       defaultState = MyState(earliest, earliest.plus(5, MINUTES)),
